VM bugfixes.
Add set command to the language, for gauge types.  Fix the compiler to create
metrics of appropriate types, warning if operations change semantics of metric.

Fix the parser to build builtin nodes with immediate list of arguments, instead
of having another intermediate list.

Add better error reporting for compile-time and runtime errors.

Uncomment lexer and parser tests that really are supported.

Wrap metrics read and write ops with a mutex.