Remove the readPartial method.
This method was only ever called from within readFile, and added an
extra loop and buffer allocation.  Instead, allocate the buffer once at
a capacity of 4096 and initial size 0, and use the returned number of
bytes read to cap the size of it.

EOF means different things to different callers, so pass it back up and
handle in the caller.