Add context from the parser state to the lexer to enable div.
Lex DIV tokens as DIV regardless of state, but in the parser indicate
when we think we're entering a regex and signal that back to the lexer
so it can jump into regex mode instead of regular mode.

Update some tests that are affected by the change in tokenisation and
source positions.