Add a NL token for lexing newlines, and use them to terminate statements.
All statements must now end with a newline.  This allows us to determine
the difference between a cond starting with a pattern expression, and a
DIV operator following an expression statement.

Future changes may alter the grammar to remove the need for the final
newline inside a statement list when terminated with an RCURLY, and
EOF.  Another option is to allow multiline expressions iff the operator
precedes the newline (multiline regexes are explicitly allowed.)

This change also ensures passing tests by adding newline tokens and
characters where necessary to make them follow the new rule.