Downgrade an INFO log to v1 so that it's less spammy.
The check for presence to handle a log update is called even for
unwatched files in a directory, so a common case is to not find files in
log paths.  Downgrading the message makes it less spammy when enabling
logs via --logtostderr.

A better fix will be to test presence before calling the method, or to
manage the directory watches and file watches separately.

Fixes: #13.