Installs goyacc from the golang tools source before starting the build.
Adds an install_build_deps target to the Makefile, and adds it as a
dependency to vm/parser.go.  Updates the path to goyacc in compiler.go
to generate the code.  Updates the commandline to execute go generate in
the Makefile to be just enough verbosity to show what is executed, and
to remove an unnecessary command.