Exercises the tailer in the example test and benchmarks.
Exits the tailer if EOF is read in oneShot mode for the example test to
terminate.

As a side effect, the benchmarks now truly measure bytes per second.  The
transfer is done via the filesystem.  I don't know for certain that the
benchmark would work correctly if fsnotify is not detecting changes in the
afero MemMapFs filesystem in between iterations.

The mtail.Serve() method needs a cleanup.