Fixes a bug with two readers on one events channel.
Both tailer and loader were trying to read all filesystem notify events from
the same channel.  This changes Events() to return a new channel and updates
the dispatcher to send each Event to every channel, so both have a chance of
acting on the event.  Prior to this, a loader may eat a log update message and
prevent mtail from timely updates to the metrics.