Typecheck patternConstNodes.
Replaces regexNode with patternConstNode.  Because patternConstNodes are
concatenable with operator plus, we now report their errors in the condNode
that they're only allowed to exist in.

Adds support for named pattern expressions and detecting when they're not
defined.  Introduces a new object class 'PatternSymbol' to do so.

Binary expressions that concatenate patterns not yet supported.

Moves the support for handling the range of a patternConstExpr to include the
quoting slashes from pattern_expr to regex_pattern in the parser.

Adds some extra verbose logs in -v2 for the lexer and parser.y vmodules.