Fix logic bug when opening a new log path.
The comment indicates we have never seen this log before, but we tell
openLogPath to treat it like we have; this bug looks like it was introduced in
ccf6a8fe6b525997e96c6b2d4db1fc701fe8cc6c in order to handle new and
not-yet-existent log files and never noticed until now.

Fixes #120.