Update the docs to point to references about GOPATH.
The documentation was not very helpful for newcomers who didn't know what the
magic keywords are to set up a GOPATH workspace.  So let's just make some
hyperlinks.

While we're at it, expand on the build docs some more, instructions how to run
the tests, and give an opinionated TLDR on first time setup.