Report a compile error when symbols are declared and not used.
Except for capture group references, which are not always controlled by the
user -- for example decorators and pattern constants can define capture groups
not intended for use by the user.

Refactor the symbol table to accept aliases for the same symbol, though, and
use that when creating named capture group references so we know when a capture
group is referred to by either name or number.

Update the one test that has a new error message.