Reorder some loop breakers to try to beat subtle timing bugs.
This test sometimes issues an extra WriteString; an alternate future has this
test fixed by allowing numbers greater than or equal to 10 to pass, but in this
timeline we try hard to figure out why the thing is being nondeterministic, and
failing.  Putting the loop break at the start of the next tick seems to win,
while testing immeidately after the increment fails as much as before.
Probably we do not wait for the loader to report completion when performing a
mtail.Close() ?