The watcher API now sends events to the registerer of the file.
This will reduce work and in some tests false positives by making sure that the
program events go to the loader, and the log events go to the tailer.

Watch receivers must now pass back a handle when registering new watches.  We
can't avoid storing the list of event channels as well as a mapping of
pathnames to channels because we need to ensure that the channels are closed by
the watcher at shutdown.  Likewise there's no point passing in channels from
the mtailserver setup and having that manage their lifetime.

Clean up the mcp_test which was not really exercising the MCP startup.