Attempt to avoid strange type assignment for string counters.
Force Text metrics to be of type String before type checking.

Counter, Gauge and Timer metrics are marked as variable, though they should be
Numeric.  Setting them to Int or Float causes type assignment to be too
specific too early.  As a side effect we can make Gauge metrics only like Floats
and Counters only like Ints, but this could be a breaking change.

Another version looks to type coercion in ASSIGN,ADD_ASSIGN case in the type
checker, attempting to work around lack of Numeric types.  We also do not
correctly encode the rT <= lT relationship in the type system for this case.