Do nonblocking opens on files.
When a named pipe is opened, it normally blocks until the other end of the pipe
is also opened.  This prevents mtail from starting up correctly.

Instead, open all files in nonblocking mode.  This means the pipe can be opened
and later reads will work if there's data available.  Opening regular files in
nonblocking mode has no real effect for our purposes, so no need to turn it
off.

We track that a file is "regular" now so we know if we have to do a truncation
test.  Pipes can't be truncated, so don't try.