Let the user disable fsnotify.
When the update rate of log files is sufficiently high, the fsnotify event
queue can overrun because mtail is not processing the queue fast enough.

(It's probable that because mtail keeps reading until EOF on each file in the
tailer that it both starves the queue processor and also obsoletes messages on
the queue.)

fsnotify does not handle this error explicitly; if a user process gets this
error then the process must assume it has lost notifications and reconstruct
all watches.  (Based on https://lwn.net/Articles/605128/)

This change makes the assumption that in such cases, fsnotify is not helping,
and explicit polling will do a better job.

Issue #195