Refactor coverage so it's rebuilt when dependencies change.
Split up each package coverage to allow for parallelism.  We need secondary
execution so the static pattern's use of wildcard gets expanded in the correct
place.  No '/' in the pattern as make doesn't like it.

Also, remove gover as it's not necessary, we can get away with a simple grep to
concatenate our cover profiles.