Error out if extra arguments are present after flag parsing is complete.
I suppose it's not obvious that, if you were to space separate some files, then
they would be ignored.

(An alternate, but behaviour changing option is to treat every extra argument
as a filename pattern to tail.  This might be a good idea, and I'm going to
think about it.)

Issue #229