Don't read from files during bootstrap, because they can block.
If we have an input that is being written to fast enough, and never has an EOF,
we can block from exiting this section and the mtail HTTP port never comes up.
This is easy to reproduce with a busy server, rsyslog, and a named pipe as
input.

But tests assume that this works so we allow a file read during startup if
oneShot mode is enabled.

Issue: #281.