Use codecov instead of coveralls.
We can't collate the coverage profiles from different runs because there's no
shared storage to write to.  We can't use a cache because there's a race
condition opening it unless we serialise each version image, and we can't use
the HTTP Cache because there's no way to know what the key is per image without
duplication.