Create a streaming, goroutine-per-log tailer implementation.
Each LogStream has a goroutine per log source which has a lifetime
that matches the source.  For files, when a rotation occurs, a new goroutine is
created.  All normal EOFs or read timeouts turn into a "poll sleep" that can be
interrupted by a context cancellation.

We cannot create a socket under test with this API as it's trying to
stat the object to find out what it is, but sockets must be created by the
listener, which in this case is mtail.  We'll need to add some magic to parse
socket names in the log patterns.

IsComplete can be used by the Tailer to determine if a LogStream has finished
work.  A LogStream may finish for various reasons (file not found, named pipe
is closed) but the assumption is that the stream cannot recover from this and
has declared itself done.  The Tailer can recreate a new LogStream on the same
pathname in the future if it so chooses.

Logstreams increment the waitgroup before starting the streaming goroutine,
allowing us to count them for correct shutdown.

The Stop method lets the Tailer stop each stream gracefully.  For example, the
Tailer needs to stop a stream for OneShot mode, or if a file no longer exists
on the filesystem during a Poll.

Issue #339, issue #289