Test for filestream rotation.
We need a better way to communicate Poll across multiple anonymous goroutines
so this method tries to resend a poll event if anyone else is listening at the
moment.

Also we don't pause for Poll if we're not at EOF anymore. Don't need to count
totalBytes anymore as we're never returning from this loop until the goroutine
is done, which also means we can exit when finished and avoid tracking it in
another boolean flag.

Also fix detection of fd change to use our newly stated FileInfo or else we use
stale values that never change and falsely trick us to thinking that we've
truncated.

Too many logs emitted here, though, they need to be downgraded to V(2).