Add a test for socketstream.
Learn that we cannot create a socket under test with this API as it's trying to
stat the object to find out what it is, but sockets must be created by the
listener, which in this case is mtail.  We'll need to add some magic to parse
socket names in the log patterns.