Add IsFinished to the LogStream interface.
IsFinished can be used by the Tailer to determine if a LogStream has finished
work.  A LogStream may finish for various reasons (file not found, named pipe
is closed) but the assumption is that the stream cannot recover from this and
has declared itself done.  The Tailer can recreate a new LogStream on the same
pathname in the future if it so chooses.