Create a Waker that wakes up idle logstreams.
Using a time.Ticker directly makes it difficult to synchronise tests, and makes
tests be time dependent.

This interface lets us create a device that in production looks like a Ticker,
but in test lets us inject events.  The testWaker also allows us to act like a
barrier to ensure test points are correctly synchronised, i.e avoiding flakes.

The production waker also takes a Context so it can be cancelled.