Create a Waker that wakes up idle logstreams.
The previous design made it difficult to test as we needed to be able to
synchronise logstreams to ensure test invariants remained true.  But in
the production use case was exactly opposite, the wake shouldn't interfere with
the flow of the stream and certainly shouldn't be forcing streams to block or
wake up too soon.

This interface lets us inject a test waker when in test, and the production
waker does all the timing itself.