Move the log pattern polling to the Tailer.
The tailer is the natural place for this method, and like previous commits here
we can ensure it's configured and running at construction time, rather than
having to set it up later in mtail's initialisation.

Also plumb through some new Options for the tailer and mtail server, and clean up
some comment style.