Make new test compile, but not work.
Untag this test as integration so that it runs with the rest of the tests.
Then, learn why it's got a compile error in it.. then learn why it
fails... Attempt to fix the use of a root-only filename but fail, and so in the
meantime mark it for skip.