Mark golangci-lint continue on error.
This is in the step, not the job, per discussion in
https://github.com/actions/toolkit/issues/399 which indicates that a
continue-on-error at the job level would cause the icon to still be a cross,
but at the step level is totally ignored.  Right now golangci-lint doesn't work
for mtail due to https://github.com/golangci/golangci-lint/issues/1559 so let's
leave it running but ignore it totally with the former setup.