Initialise flag global in test to avoid side effects.
The flag value isn't initialised because we never call flag.Parse, and
testutil.TestSetFlag doesn't seem to reset the value as a result -- probably
the value is nil?

Anyway, manually setting the value at the start of the test seems like a nice
hack.  Really we should avoid the flag value here and pass it in to the
pushOptions somehow.