Replace Log with Logstream.
This change is pretty invasive as lots of the tests depended on side effects
from Log's implementation.

This adds a field to TestServer to let us specify the number of wakers to wait
on during PollWatched, and in some cases we remove unnecessary tests like in
log_glob_test where we don't need to count lines, just logs.

The timeout is raised slightly because of the -cpu flag which makes examples
sometimes run longer than a minute.

The watcher is removed from the tailer here, as we no longer need it, replaced
by the logstream poll waker which we pass to the logstreams.

Some expvars have changed their names.

TestServer only wakes logstreams, the other things (LogPatternPoll and
StaleLogGc) are handled directly by the PollWatched method.

TestServer Poll ordering changed so that logstreams just created can get waited
on, and context cancellation ordering fixed in shutdown.

The parameter 'seekToStart' in logstream.New is changed to 'streamFromStart'
which makes more sense about what it's doing.

Filestream's handling of notexist errors are fixed -- i.e. it doesn't need to,
the manager (tailer) will fix it.

Ooh so much code deleted, so good.

Fixes: #339, #249, #289