Convert the Loader to use a logline channel.
Previously we had used a channel to pass loglines through mtail.  That was
changed to attempt to use trace spans to measure the time spent dealing with
each logline and making cool flamegraphs.  But that pattern makes it more
difficult to shutdown, so moving it back to the channel producer/consumer
model.

Here the loader sets up each VM with a lines channel, and uses that to signal
shutdown to each VM.  The loader is responsible for waiting for all VM Run
goroutines to finish.