Shutdown the loader entirely with just the producer channel close.
Loader now just takes the lines channel and will perform its own shutdown (via
close()) when the channel ends.  In order to synchronise shutdown, Close()
remains with a waitgroup; this ought to change to a provided waitgroup from
mtail.

mtail takes on the role of the logline.Processor interface temporarily, until
Tailer no longer takes a logline.Processor parameter.

It's important to close the lines channel before waiting for Close in the unit
tests or deadlock occurs.