Shutdown the tailer with context cancellation.
Manage the tailre startup and shutdown neatly by using the mtail waitgroup to
know when it's done, and use another internal waitgorup to wait for tailer's
goroutines.  Add a special closer goroutine that manages the lines channel and
internal routine shutdown.

Return a new func from the makeTestTail method that triggers a tailer shutdown.