Simplify mtail server shutdown.
Move the SIGINT/TERM handler to the commandline frontend, and let it cancel the
context.

Avoid the need for a Close method, which removes some mess around concurrent
shutdown.

HTTP server startup and shutdown is all done inside Serve() now.  We don't need
WaitForShutdown as it's signalled via the context.

Increase the shutdown timeout because for some reason it takes 6 seconds on
average, on my machine, under test.  This might be because the server startup
test is polling the HTTP server.  It's a TODO to change that test helper.

We need to keep mtail's cancel context around so we can ensure cancellation
under test when Run() exits.