Modernise exporter init and shutdown.
Use a context and waitgroup for signalling shutdown to the exporter, instead of
dangling goroutines around.  Move a commandline flag out of the package and up
to the main program.  Still a bit of a mess but an improvement.  I dislike
these exporters but I don't know if anyone is using collectd/statds/graphite in
production with mtail.