Remove the test server startup poll and restore graceful shutdown.
Graceful shutdown took a long time because we'd created a connection during
test, and the behaviour of net/http.Server.Shutdown is to poll every 500ms to
clean up.  Presumably there's a 5s timeout in the net.Listener as well for idle
connections because the tests always seem to take at least that long to stop.

Removing the test setup poll makes the server shutdown faster; kill two TODOs
in one commit!