Hash the program content so that they are not reloaded every time.
Keep a SHA256 hash of the program content so when we go to reload the programs,
we only reload the ones that have changed.  This makes tests less spammy and
confusing (the old vms exit every time we PollWatched!) and also probably will make
production less confusing for the same reason.