Replace Log with Logstream.
This change is pretty invasive as lots of the tests depended on side effects
from Log's implementation.

This adds a field to TestServer to let us specify the number of wakers to wait
on during PollWatched, and in some cases we remove unnecessary tests like in
log_glob_test where we don't need to count lines, just logs.

The timeout is raised slightly because of the -cpu flag which makes examples
sometimes run longer than a minute.

The watcher is removed from the tailer here, as we no longer need it, replaced
by the logstream poll waker which we pass to the logstreams.

Tailer cannot await context.Done in tests and oneshot mode, though.

Some expvars have changed their names.

Fixes #339, fixes #289