Change the testwaker barrier function to know how many to wait on.
The number of goroutines we need to wait on is not constant, so provide it
during the call to awaken and instead of a waitgroup, use a token channel to
count the wakees arriving.

In one instance, the truncation test, we have a race condition where the test
runs ahead of the filestream routine, so occasionally the line is not read when
we come to truncate.  This tricks the filestream into thinking the offset has
not changed.  In the Litle Book of Semaphores
https://greenteapress.com/wp/semaphores/ and Marecelo Bytes'
https://blog.ksub.org/bytes/2015/11/08/reusable-barriers/ Go version, we learn
the correct method is to barrier the test on the goroutine returning to Wake(),
but this is more tricky because we need to know how many filestreams will be
active *next*.

Instead in this one instance we can just do the barrier twice to ensure that we
get one pass through the filestream and know the line has been read, decoded,
and sent before we continue in the test.