Add log deletion handling.
Unskip the delete test and add a poll routine that checks the existing log
streams against the filesystem.  If any are gone, stop the stream and clean up.

Count opens instead of rotations, though.  If we count log opens, rather than
rotations, then we can correctly test for a race during log rotation and soft
link change without resorting to terrible tricks between tailer and filestream.

Specifically the problem is that during the race, we cannot tell the difference
between a delete and a partially executed rename.  The result is that the
tailer attempts to remove the filestream, but the filestream was attepmting to
count a rotation.

If instead we just let the delete occur, we can just count the number of
successful open calls and both paths have the same result.

We also now need to stop trying to delete in Tailer, and let the Filestream
deal with it.  The Tailer will race with the filestream when it calls Stop,
which is not correct because it may interfere with a rotation-created
goroutine (i.e. stopping all log source routines in a single stream), or the
calls to Stop and the filestream loop are not synchronised.

This means we need to turn the testWaker into a proper barrier and instead of
letting the test and filestream run alongside, wait for the filestream to
return to Wake so we know it's waiting for the test to perform the next
action.  This looks very messy so I hope to be able to clean it up in the future.

Fixes #338, fixes #300