Correctly identify negative integers as Int in capture groups.
As a side effect we end up removing a type cast dance in a few cases.

Also factor out the syntax.Parse call to ensure we always call it with the
correct flags -- in our case we go for the Perl/Python RE2 compatibility which
is at the moment identical to how Go's regexp.Compile behaves.

Fixes #332