Fix logic of how pipeStream is shutdown.
Pipes by nature do not have an end until the pipe is closed, so they do not
behave like a filestream and thus should not be treated the same as a file as
far as EOF and timeouts go.

Thus, we do not need a Stop signal to tell a pipestream when to gracefully end
reading a pipe -- when the pipe is closed is the correct time to gracefully
end.  A stop channel and a pipe closure are the same logical event so we do not
need the former.

Rewrite the file/pipe comparison test to avoid test races due to trying to call
testing.Fatal during setup.

Also rewrite the pipe tests to simulate correct pipe behaviour -- write only in
test, blocking mode.

Fixes #425