Benchmark adds to the metrics store.
Add use of testing.quick.Value to autogenerate random metrics.  testing.quick
does not look like its' very loved - bugs like
https://github.com/golang/go/issues/21745 make it look like it'll never get
love, and
https://stackoverflow.com/questions/38906954/generate-a-tree-of-structs-with-testing-quick-respecting-invariants
would be really useful if the suggestion for recursiveness was implemented.
The API is clunky to use without good examples as well; reflect.Value and
TypeOf is confusing!

The best example of how to autogenerate records is in quick.sizedValue():
https://golang.org/src/testing/quick/quick.go#L66

The rest of the benchmark is copied from the standard library
sync/map_bench_test.go and updated to use table driven style.

Issue #341