Switch the store RWMutex to a pair of categorical mutices.
The post in The Little Go Book of Semaphores
https://blog.ksub.org/bytes/2016/04/10/the-search-insert-delete-problem/ is a
fascinating read -- well, the whole thing is really -- and suggests a solution
to the general case here.  I think this is actually overkill -- we don't delete
anyuthing from the store except in initialisation (which is probably also
unnecessary) so we don't need the inserter lock as well -- the problem is
probably just that Add took the wrong lock during the search part.  However it
was fun to figure out how to apply this pattern and if we can't have fun, why
are we even here?

Benchmarks shows a massive reduction in time for the new Store/Add and
Store/Iterate families.

Issue #341