Export histogram buckets and total count for Graphite
Up until now when exporting histograms for Graphite only the sum of the
values was being exported as one metric. From now on the counts of
values in distinct buckets along with the total count is exported too,
as separate metrics, so histograms are now usable in Graphite too.

Perhaps similar code should be added to collectd, statsd and varz
metricTo* converters too...