Correct TestTestWakerTwoWakeups by waiting for the actual wakeup.
There was a race where we'd skip ahead in the `Wake()` and accidentally wait on
the same wake channel a second time, which would get the wrong result when that
occurs because it would have been `awaken()`d by then.  Forcing the first
iteration to block on `Wake()` ensures the second one gets a new channel.