Refactor the PatternExpr check and compile to emit Match safely.
This change wraps an implicit match expression as a UnaryExpr with a Match
operator so we can tell it apart from a raw PatternExpr as an argument to a
function.

As a side effect it makes the codegen neater by not having to rewrite
instructions, because we can do the right thing the first time.