Simplify pipestream by correctly interrupting a read on context cancellation.
https://github.com/golang/go/issues/20280#issuecomment-655588450 has an
excellent example on how to interrupt a read on context cancellation, which
means we don't need to set a read deadline on every read attempt.  Then, the
error handling is simplified, and as a side effect we aren't having spurious
wakeups on idle fifos.

This also means we no longer have a race on first read when the write end of a
fifo isn't ready.

Issue: #486