Avoid a race by opening the write end first.
Use O_WRONLY to open the fifo in test means we ensure we have a
non-EOF-signalling pipe for the pipestream when it gets to fd.Read().  This is
possible if the scheduler schedules the new goroutine before the test proceeds.

Not convinced that there isn't still a race in production.