Tidy up the datagram stream.
Raise the read buffer size to be a magic number that looks like the default
Linux datagram buffer size.  If we don't read the entire datagram in one go, we
will end up discarding bytes.

We fake Stop/EOF by setting a read deadline on Stop to allow us to do a second
read and then terminate if no new bytes show up in that time.