Unify returns a Type.
As described, Unify returns the unified type, which helps when unifying
Operators (and can't replace either due to how the call works.)  Finally get to
remove the commented-out SetInstance which couldn't work, with the correct `return`.

It looks like we uncovered some bugs with Equals not testing Variables for
equality correctly, and the TypeUnification had never actually tested against
`expected` before.

This also redoes Operator-Operator unification to construct a return Operator
from both the Name and Args unification branches.