Don't run a second goroutine on the logstreamWaker.
Test synchronisation relies on the logstreams being the only wakee waiting on
the testwaker.  If we use it in a second stream, we either have to bump the
awaken counter or remove it.

The PollLogstream method was incorrectly named; renaming for more clarity means
we see that we don't need to call it on the same waker but instead clean up
streams when we poll for new patterns, less frequently but frequent enough.

This undoes some of the changes from #603 as mentioned in that comment thread.

Fix up names and docstrings that were confusing.