Avoid test hang when net.Conn.Write times out
As described in #614, some times writing a zero-length message fails with a
time out error, causing the goroutine to fail without marking the stream as
finished, and so the test server never finishes.

This commit adds a work-around by retrying the `Write` call until it succeeds or
returns a non-timeout error. It also adds a `Sleep` call between tries as
otherwise this becomes a spin loop, filling the log with timeout errors.