Fix memory leak on corrupt unicode character
See the added test.
When a corrupt unicode character is received, mtail did not process
the following logs but kept them in memory.

With this commit RuneErrors on parsing unicode characters are ignored
when there are enough characters after it in the buffer.
This keeps the functionality intact that bytes that gave a RuneError are
being reparsed, but with a larger buffer, to not break unicode
characters which were split in half from the buffered reading on the
logfile being read.

The issue was introduced in v3.0.0-rc50