Explicitly stop the `Exporter` when main routines are done.
The `Exporter` and `HTTPServer` are not connected by channels to the `Runtime`, only
indirectly through mutating effects on the Store.  As such we can't rely on the
`Exporter` to shut down automatically when the `Runtime` shuts down.

Here instead we remove the `WaitGroup` and allow an `Exporter` to be shut down
explicitly after the `WaitGroup` is `Done`.

Issue: #331