refactor: Immediately tail sockets when discovered.
Instead of collecting sockets and tailing them after, just tail them straight
away.

This makes the "early exit" check for "no things" broken, so replace that with
a check in the final shutdown handler that there were no active globs.  This
will be replaced in the next few changes once migrated to goroutines and we can
use only the `WaitGroup` for counting globs, too, but in the meantime this is
more correct than the previous code.

Issue: #331