fix: Exit a logstream at EOF when OneShotEnabled.
This means a stream will shut itself down and does not need the tailer to
instruct it to do so.  It thus avoids a race when the tailer doesn't know if
the stream has performed enough reads.

We need to separate the meaning of oneShot mode and streamFromStart as the
former tells us how to exit a stream, and the latter tells us how to start
reading a new file after log rotation.