fix: fix symbol checking and creation for ``subst`` with a regexp
This change updates the ``checkRegex`` helper to not track symbols for a
regexp when it is used by the builtin ``subst``. In this case the VM
will not store these matches, and therefore these symbols will not be
available to the program. The compiler will, however, compile these
regexp and add them to the regexp table for use by the VM, and multiple
calls to ``subst`` are already handled by the VM.

fixes: https://github.com/google/mtail/issues/693, https://github.com/google/mtail/issues/810