ci: Run golangci-lint on pull requests and update version.
Attempt to fix the permissions to write annotations if it can.  Annotations are
not comments, but it's better than nothing, and thus having to find the check
failure and read the file when there's an error.

When we're on a PR, enable only new issues.  PR writers don't care for old problems.

Upgrade the version while we're here.